+++ packages/gestalt/src/LegacyController.js
@@ -17,7 +17,7 @@ const SIZE_WIDTH_MAP = {
 type OwnProps = {
   accessibilityLabel?: string,
   anchor: HTMLElement,
-  bgColor: 'blue' | 'darkGray' | 'orange' | 'red' | 'white',
+  bgColor: 'blue' | 'darkGray' | 'red' | 'white',
   border?: boolean,
   caret?: boolean,
   children?: ReactNode,