+++ packages/gestalt/src/index.js
@@ -5,6 +5,7 @@ import ActivationCard from './ActivationCard';
 import Avatar from './Avatar';
 import AvatarGroup from './AvatarGroup';
 import Badge from './Badge';
+import BannerCallout from './BannerCallout';
 import BannerOverlay from './BannerOverlay';
 import BannerSlim from './BannerSlim';
 import BannerUpsell from './BannerUpsell';
@@ -12,7 +13,6 @@ import Box from './Box';
 import Button from './Button';
 import ButtonGroup from './ButtonGroup';
 import ButtonLink from './ButtonLink';
-import Callout from './Callout';
 import Checkbox from './Checkbox';
 import Collage from './Collage';
 import Column from './Column';
@@ -91,6 +91,7 @@ export {
   Avatar,
   AvatarGroup,
   Badge,
+  BannerCallout,
   BannerOverlay,
   BannerSlim,
   BannerUpsell,
@@ -98,7 +99,6 @@ export {
   Button,
   ButtonGroup,
   ButtonLink,
-  Callout,
   Checkbox,
   Collage,
   ColorSchemeProvider,